cat_dog_with_embedding/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
transformers==4.35.0
sentence-transformers==2.2.2
qrcode==7.4.2
python-multipart==0.0.6
python-decouple==3.8
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9 